chore: move escaping to util file
