+++ src/main/kotlin/io/voidx/html/util/HtmlUtil.kt
@@ -1,0 +1,14 @@
+package io.voidx.html.util
+
+fun escapeHtmlAttr(value: String): String =
+    value
+        .replace("&", "&amp;")
+        .replace("\"", "&quot;")
+        .replace("<", "&lt;")
+        .replace(">", "&gt;")
+
+fun escapeHtmlText(value: String): String =
+    value
+        .replace("&", "&amp;")
+        .replace("<", "&lt;")
+        .replace(">", "&gt;")
